added typescript to the backend

+++ api/src/controller/userController.ts
@@ -1,7 +1,9 @@
-import { UserModel } from "../model/User.js";
+import { UserModel } from "../model/UserModel.js";
 import { deleteCloudnaryImage } from "./imageController.js";
+import { Request, RequestHandler, Response } from "express";
+import { RequestWithUser } from "../model/UserModel.js";
 
-export const createUser = async (req, res) => {
+export const createUser = async (req: Request, res: Response) => {
   const user = await UserModel.create(req.body);
 
   res.status(200).json({
@@ -10,8 +12,9 @@ export const createUser = async (req, res) => {
   });
 };
 
-export const handleMeRoute = async (req, res) => {
+export const handleMeRoute: RequestHandler = async (req: RequestWithUser, res: Response) => {
   try {
+    if (!req.user) throw Error;
     const user = await UserModel.findById(req.user._id).select("-__v");
 
     res.status(200).json({
@@ -26,14 +29,15 @@ export const handleMeRoute = async (req, res) => {
   }
 };
 
-export const updateUser = async (req, res) => {
+export const updateUser = async (req: RequestWithUser, res: Response) => {
   try {
-    if (req.body.profilePhoto) {
+    if (req.body.profilePhoto && req.user) {
       const user = await UserModel.findById(req.user._id);
-      if (user.profilePhoto) {
+      if (user?.profilePhoto) {
         deleteCloudnaryImage(user.profilePhoto.publicId);
       }
     }
+    if (!req.user) throw Error;
     const user = await UserModel.findByIdAndUpdate(req.user._id, { ...req.body }, { new: true });
 
     res.status(200).json({
